Bluetooth 방향 종료 전송 코드 리팩토링
